requirements.txt
livekit-agents~=0.8
livekit-plugins-deepgram~=0.6
livekit-plugins-openai~=0.7
livekit-plugins-silero~=0.6

python-dotenv~=1.0